feat: initial commit
Start off December 2021 right; get your Advent of Code problems ready to
be solved.

This year, we're in Elixir. I'm not terribly familiar with the language,
so prepare for a lot of documentation diving and problem solving
on-the-fly. Hopefully, people will help find better solutions than what
I've put together as I go.